refactor: descripcion de boton y le seteamos el estilo que trae por defecto de google
